requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
Pillow
joblib
python-multipart